+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - `head` package now has `GetStamp` method to find a stamp by its provider from an array.
 - `num.Percentage` has `Base()` method to access base amount.
 
+### Change
+
+- Renaming `mx.StampProviderSATUUID` constant to just `mx.StampSATUUID`.
+
 ### Fixed
 
 - UUID `IsNotZero` will not raise error for empty UUIDs.